fixing updates
